% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardiac_atlas}
\alias{cardiac_atlas}
\title{Cardiac Atlas Data}
\format{
A named \code{list} of different cardiac views, each with their underlying
\code{shapefiles}. All of these shapes are a combination of line and polygon
geometries. The types are the following: A
\itemize{
\item \strong{extent} = map background that the shapes are built upon
\item \strong{chambers} = individual anatomical chambers within the heart
\item \strong{conduction} = elements of the conduction system
\item \strong{projections} = involutions and foldings that compose critical tissue
within the chamber cavities
\item \strong{septum} = dividing structures between chambers
\item \strong{valves} = cardiac valves (tricuspid, mitral/bicuspid, aortic,
pulmonic)
\item \strong{vessels} = venous and arterial structures
}

The current projections that are available under the \strong{cardiac_atlas}
object are:
\itemize{
\item \strong{right_atrium_rao} = right anterior oblique view focused on the right
atrium
}
}
\source{
Created by author using the \href{https://qgis.org/en/site/}{QGIS}
software with templates created from the "Atlas of Cardiac Anatomy" by Mori
and Shivkumar (2022).
}
\usage{
cardiac_atlas
}
\description{
A series of \code{shapefile} data that represent cardiac anatomy in 2D.
}
\author{
Anish S. Shah, MD/MS
}
\keyword{datasets}
