% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_resnick.R
\name{change_dim_br_ll}
\alias{change_dim_br_ll}
\title{Log likelihood of spectral density of Brown-Resnick process with changing dimentions. Relies of function br_ll()}
\usage{
change_dim_br_ll(pars, dt, lcs, vr, conditioned.site = 1)
}
\arguments{
\item{pars}{(numeric) Parameter values of variogram which parameterises the Brown-Resnick process.}

\item{dt}{(list) Observations of extreme process, with standarised margins.}

\item{lcs}{(list) List of matrices that describe locations of observations in corresponding position in list dt.}

\item{vr}{(numeric) Variogram function, evaluates with "pars"}

\item{conditioned.site}{(numeric) Site that is "conditioned" on.}
}
\value{
(numeric) Log-liklihood
}
\description{
Log likelihood of spectral density of Brown-Resnick process with changing dimentions. Relies of function br_ll()
}
