% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{dgev}
\alias{dgev}
\title{Density function of the generalised extreme value distribution}
\usage{
dgev(x, loc, scale, shape, log = F)
}
\arguments{
\item{x}{(numeric) data}

\item{loc}{(numeric)}

\item{scale}{(numeric)}

\item{shape}{(numeric)}

\item{log}{(logical) If true, returns log}
}
\value{
(numeric)
}
\description{
Density function of the generalised extreme value distribution
}
\examples{
dgev(x = 0.5, loc = 0, scale = 1, shape = 0)
}
