% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{dgp}
\alias{dgp}
\title{Density function of the generalised Pareto distribution}
\usage{
dgp(x, scale, shape, log = F)
}
\arguments{
\item{x}{(numeric) data}

\item{scale}{(numeric)}

\item{shape}{(numeric)}

\item{log}{(logical) If true, returns log}
}
\value{
(numeric)
}
\description{
Density function of the generalised Pareto distribution
}
\examples{
dgp(x = 0.5, scale = 1, shape = 0)
}
