% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{dhr}
\alias{dhr}
\title{Density function of the bi-variate Hüsler-Reiss copula}
\usage{
dhr(u, v, lambda, log = F)
}
\arguments{
\item{u}{(numeric) Uniform margin of copula}

\item{v}{(numeric) Uniform margin of copula}

\item{lambda}{(numeric) Dependence parameter of the Hüsler-Reiss copula}

\item{log}{(logical) If true returns log}
}
\value{
(numeric)
}
\description{
Density function of the bi-variate Hüsler-Reiss copula
}
\examples{
dhr(u = c(0.1, 0.2, 0.4), v = c(0.6, 0.7, 0.2), lambda = 1)
}
