% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploritory.R
\name{emp_chi}
\alias{emp_chi}
\title{Internal function used to estimate emprical pairwise extremal dependance coefficient}
\usage{
emp_chi(dat, u, data_scale = T, lower_diag = T)
}
\arguments{
\item{dat}{(data.frame or matrix) Each column corresponds to a different site/variable, with observations in each row.}

\item{u}{(numeric) Quantile at which to estimate extremal dependence coefficient}

\item{data_scale}{If TRUE, the default, function transforms the data to uniform margins using an empirical ranking.}

\item{lower_diag}{If TRUE, the default, plot lower diagonal elements of extremal coefficient (chi) matrix}
}
\description{
Internal function used to estimate emprical pairwise extremal dependance coefficient
}
