% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploritory.R
\name{explore_chi}
\alias{explore_chi}
\title{Estimate and plot matrix pairwise dependence measure}
\usage{
explore_chi(dat, lower_diag = T, lims = NA)
}
\arguments{
\item{dat}{(data.frame or matrix) Each column corresponds to a different site/variable, with component-wise maxima in each row.}

\item{lower_diag}{If TRUE, the default, plot lower diagonal elements of extremal coefficient (chi) matrix}

\item{lims}{(numeric) Vector of length 2 controlling the limits of the scale for plotting extremal coefficient. Default plot has limit c(0,1).}
}
\value{
(ggplot2::ggplot) Plotted extremal coefficient (chi) matrix
}
\description{
This function plots the estimated extremal coefficient (chi) matrix derived from an estimated dependence matrix of the Hüsler-Reiss copula where data is on any scale. The function transforms the data to uniform margins using an empirical ranking.
}
