% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploritory.R
\name{explore_emp_chi}
\alias{explore_emp_chi}
\title{This function plots the extremal coefficient (chi) matrix estimated empirically.}
\usage{
explore_emp_chi(dat, u = 0.5, data_scale = T, lower_diag = T, lims = NA)
}
\arguments{
\item{dat}{(data.frame or matrix) Each column corresponds to a different site/variable, each row is an observation.}

\item{u}{(numeric) Quantile at which to estimate extremal dependence coefficient}

\item{data_scale}{If TRUE, the default, function transforms the data to uniform margins using an empirical ranking.}

\item{lower_diag}{If TRUE, the default, plot lower diagonal elements of extremal coefficient (chi) matrix}

\item{lims}{(numeric) Vector of length 2 controlling the limits of the scale for plotting extremal coefficient. Default plot has limit c(0,1).}
}
\value{
(ggplot2::ggplot) Plotted extremal coefficient (chi) matrix
}
\description{
This function plots the extremal coefficient (chi) matrix estimated empirically.
}
