% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploritory.R
\name{explore_lambda}
\alias{explore_lambda}
\title{This function plots the estimated dependence matrix of the Hüsler-Reiss copula where data is on any scale. The function transforms the data to uniform margins using an empirical ranking.}
\usage{
explore_lambda(dat, lower_diag = T)
}
\arguments{
\item{dat}{(data.frame or matrix) Each column corresponds to a different site/variable, with component-wise maxima in each row.}

\item{lower_diag}{If TRUE, the default, plot lower diagonal elements of dependence matrix}
}
\value{
(ggplot2::ggplot) Plotted dependence matrix
}
\description{
This function plots the estimated dependence matrix of the Hüsler-Reiss copula where data is on any scale. The function transforms the data to uniform margins using an empirical ranking.
}
