% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{fit_gev_bay}
\alias{fit_gev_bay}
\title{Fitting a GEV in a bayesian framework}
\usage{
fit_gev_bay(
  x,
  prior_mean_loc = 1,
  prior_sd_loc = 1,
  prior_mean_xi = 0,
  prior_sd_xi = 1,
  prior_mean_sig = 1,
  prior_sd_sig = 0,
  chains = 2,
  iter = 2000,
  warmup = floor(iter/2),
  cores = 2,
  thin = 1,
  init = 0
)
}
\arguments{
\item{x}{(numeric) data}

\item{prior_mean_loc}{(numeric) Mean of normal prior of location parameter of GEV distribution}

\item{prior_sd_loc}{(numeric) Standard deviation of normal prior of location parameter of GEV distribution}

\item{prior_mean_xi}{(numeric) Mean of normal prior of shape parameter of GEV distribution}

\item{prior_sd_xi}{(numeric) Standard deviation of normal prior of shape parameter of GEV distribution}

\item{prior_mean_sig}{(numeric) Mean of normal prior of scale parameter of GEV distribution}

\item{prior_sd_sig}{(numeric) Standard deviation of normal prior of scale parameter of GEV distribution}

\item{chains}{(numeric) see ?rstan::sampling}

\item{iter}{(numeric) see ?rstan::sampling}

\item{warmup}{(numeric) see ?rstan::sampling}

\item{cores}{(numeric) see ?rstan::sampling}

\item{thin}{(numeric) see ?rstan::sampling}

\item{init}{(numeric) see ?rstan::init}
}
\value{
(list(estimate (numeric), post (numeric), fitted_model (stan model))) List of 3 elements,
first element is the estimate of the GEV parameters, calculated as the mean of the posterior distribution.
Second element is the posterior distribution samples. Third element is the full fitted stan model.
}
\description{
Fitting a GEV in a bayesian framework
}
\examples{
x = rgev(100, scale = 1, scale = 1, shape = -0.1)
fit_gev_bay(x)
}
