% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{fit_gevd}
\alias{fit_gevd}
\title{Fitting a generalised extreme value distribution}
\usage{
fit_gevd(x, initial_est = c(0, 1, 0))
}
\arguments{
\item{x}{(numeric) Vector of observations}

\item{initial_est}{(numeric) Initial estimates of distribution parameter (scale, shape)}
}
\value{
(list(estimate (numeric), ci (matrix))) List of 2 elements,
first element is the estimates of the generalised extreme value distribution parameters,
second element is the estimated 95% confidence interval from the Hessian, the i-th row corresponds to the i-th estimate.
}
\description{
Fitting a generalised extreme value distribution
}
\examples{
x = rgev(500, loc = 5, scale = 1, shape = -0.2)
fit_gevd(x)
}
