% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{fit_gpd}
\alias{fit_gpd}
\title{Fitting a generalised Pareto distribution}
\usage{
fit_gpd(x, initial_est = c(1, 0))
}
\arguments{
\item{x}{(numeric) Vector of observations}

\item{initial_est}{(numeric) Initial estimates of distribution parameter (scale, shape)}
}
\value{
(list(estimate (numeric), ci (matrix))) List of 2 elements,
first element is the estimates of the generalised Pareto distribution parameters,
second element is the estimated 95% confidence interval from the Hessian, the i-th row corresponds to the i-th estimate.
}
\description{
Fitting a generalised Pareto distribution
}
\examples{
x = rgp(500, scale = 1, shape = -0.3)
fit_gpd(x)
}
