% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{fit_hrc_bay}
\alias{fit_hrc_bay}
\title{Fitting a bi-variate H端sler-Reiss copula in a bayesian framework}
\usage{
fit_hrc_bay(
  u,
  v,
  prior_mean = 0,
  prior_sd = 1,
  chains = 4,
  iter = 1000,
  warmup = floor(iter/2),
  cores = 2,
  thin = 1
)
}
\arguments{
\item{u}{(numeric) Uniform margin of copula}

\item{v}{(numeric) Uniform margin of copula}

\item{prior_mean}{(numeric) Dependence parameter of the H端sler-Reiss copula}

\item{prior_sd}{(numeric) If true returns log-Likelihood, default is False}

\item{chains}{(numeric) see ?rstan::sampling}

\item{iter}{(numeric) see ?rstan::sampling}

\item{warmup}{(numeric) see ?rstan::sampling}

\item{cores}{(numeric) see ?rstan::sampling}

\item{thin}{(numeric) see ?rstan::sampling}
}
\value{
(list(estimate (numeric), post (numeric), fitted_model (stan model))) List of 3 elements,
first element is the estimate of the H端sler-Reiss dependence parameter, calculated as the mean of the posterior distribution.
Second element is the posterior distribution samples. Third element is the full fitted stan model.
}
\description{
Fitting a bi-variate H端sler-Reiss copula in a bayesian framework
}
\examples{
dat = rhr(1000, 1.2)
fit_hrc_bay(dat[,1], dat[,2], chains = 2)
}
