% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{fit_hrc_gev_bay}
\alias{fit_hrc_gev_bay}
\title{Fitting a bi-variate Hüsler-Reiss copula in a bayesian framework jointly with GEV margins estimates}
\usage{
fit_hrc_gev_bay(
  x,
  y,
  prior_mean = 1,
  prior_sd = 1,
  prior_mean_xi_1 = 0,
  prior_sd_xi_1 = 1,
  prior_mean_sig_1 = 1,
  prior_sd_sig_1 = 1,
  prior_mean_xi_2 = 0,
  prior_sd_xi_2 = 1,
  prior_mean_sig_2 = 1,
  prior_sd_sig_2 = 1,
  prior_mean_loc_1 = 1,
  prior_sd_loc_1 = 1,
  prior_mean_loc_2 = 1,
  prior_sd_loc_2 = 1,
  chains = 2,
  iter = 1000,
  warmup = floor(iter/2),
  cores = 2,
  thin = 1
)
}
\arguments{
\item{x}{(numeric) margin 1, assumed to be generalised extreme value distributed}

\item{y}{(numeric) margin 2, assumed to be generalised extreme value distributed}

\item{prior_mean}{(numeric) Mean of normal prior of dependence parameter of the Hüsler-Reiss copula}

\item{prior_sd}{(numeric) Standard deviation of normal prior of dependence parameter of the Hüsler-Reiss copula}

\item{prior_mean_xi_1}{(numeric)  Mean of normal prior on shape parameter of GPD of margin 1}

\item{prior_sd_xi_1}{(numeric)  Standard deviation of normal prior on shape parameter of GPD of margin 1}

\item{prior_mean_sig_1}{(numeric) Mean of normal prior on scale parameter of GPD of margin 1}

\item{prior_sd_sig_1}{(numeric)  Standard deviation of normal prior on scale parameter of GPD of margin 1}

\item{prior_mean_xi_2}{(numeric)  Mean of normal prior on shape parameter of GPD of margin 2}

\item{prior_sd_xi_2}{(numeric)  Standard deviation of normal prior on shape parameter of GPD of margin 2}

\item{prior_mean_sig_2}{(numeric)  Mean of normal prior on scale parameter of GPD of margin 2}

\item{prior_sd_sig_2}{(numeric)  Standard deviation of normal prior on scale parameter of GPD of margin 2}

\item{prior_mean_loc_1}{(numeric) Mean of normal prior on location parameter of GPD of margin 1}

\item{prior_sd_loc_1}{(numeric)  Standard deviation of normal prior on location parameter of GPD of margin 1}

\item{prior_mean_loc_2}{(numeric)  Mean of normal prior on location parameter of GPD of margin 2}

\item{prior_sd_loc_2}{(numeric)  Standard deviation of normal prior on location parameter of GPD of margin 2}

\item{chains}{(numeric) see ?rstan::sampling}

\item{iter}{(numeric) see ?rstan::sampling}

\item{warmup}{(numeric) see ?rstan::sampling}

\item{cores}{(numeric) see ?rstan::sampling}

\item{thin}{(numeric) see ?rstan::sampling}
}
\value{
(list(estimate (numeric), post (list), fitted_model (stan model))) List of 3 elements,
first element is the estimate of the GPD margin parameters as well as the Hüsler-Reiss dependence parameter, calculated as the mean of the posterior distribution.
Second element are the posterior distribution samples. Third element is the full fitted stan model.
}
\description{
Fitting a bi-variate Hüsler-Reiss copula in a bayesian framework jointly with GEV margins estimates
}
\examples{
# sample from a Hüsler-Reiss copula
copula_sample = rhr(1000, 2)

# Transform margins to be GPD
margin_1 = qgev(copula_sample[,1], loc = 1, scale = 1.2, shape = -0.11)
margin_2 = qgev(copula_sample[,2], loc = 1.1, scale = 1.3, shape = -0.09)

# fit margins and copula jointly
fit_hrc_gev_bay(margin_1, margin_2)
}
