% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrd-package.R
\docType{package}
\name{hrd-package}
\alias{hrd}
\alias{hrd-package}
\title{hrd: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dairer/hrd}
  \item Report bugs at \url{https://github.com/dairer/hrd/issues}
}

}
\keyword{internal}
