% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{pgev}
\alias{pgev}
\title{Cumulative distribution function of the generalised extreme value distribution}
\usage{
pgev(x, loc, scale, shape)
}
\arguments{
\item{x}{(numeric) data}

\item{loc}{(numeric)}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Cumulative distribution function of the generalised extreme value distribution
}
\examples{
pgev(x = 0.5, loc = 0, scale = 1, shape = 0)
}
