% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{pgp}
\alias{pgp}
\title{Cumulative distribution function of the generalised Pareto distribution}
\usage{
pgp(x, scale, shape)
}
\arguments{
\item{x}{(numeric) data}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Cumulative distribution function of the generalised Pareto distribution
}
\examples{
pgp(x = 0.5, scale = 1, shape = 0)
}
