% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{qgp}
\alias{qgp}
\title{Quantile function of the generalised Pareto distribution (inverse cumulative distribution function)}
\usage{
qgp(x, scale, shape)
}
\arguments{
\item{x}{(numeric) data}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Quantile function of the generalised Pareto distribution (inverse cumulative distribution function)
}
\examples{
pgp(x = 0.5, scale = 1, shape = 0)
}
