% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{rgev}
\alias{rgev}
\title{Generate random samples from the generalised extreme value distribution}
\usage{
rgev(n, loc, scale, shape)
}
\arguments{
\item{n}{(numeric) Number of samples}

\item{loc}{(numeric)}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Generate random samples from the generalised extreme value distribution
}
\examples{
rgev(n = 100, scale = 1, shape = 0)
}
