% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{rgp}
\alias{rgp}
\title{Generate random samples from the generalised Pareto distribution}
\usage{
rgp(n, scale, shape)
}
\arguments{
\item{n}{(numeric) Number of samples}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Generate random samples from the generalised Pareto distribution
}
\examples{
rgp(n = 100, scale = 1, shape = 0)
}
