% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{rhr}
\alias{rhr}
\title{Generate random samples from the bi-variate H端sler-Reiss copula}
\usage{
rhr(n, lambda)
}
\arguments{
\item{n}{(numeric) Number of random samples}

\item{lambda}{(numeric) Dependence parameter of the H端sler-Reiss copula}
}
\value{
(Matrix) Random samples from bi-variate H端sler-Reiss copula
}
\description{
Generate random samples from the bi-variate H端sler-Reiss copula
}
\examples{
rhr(n = 10, lambda = 1)
}
