% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{fit_hrc}
\alias{fit_hrc}
\title{Fitting a bi-variate Hüsler-Reiss copula}
\usage{
fit_hrc(u, v, initial_est)
}
\arguments{
\item{u}{(numeric) Uniform margin of copula}

\item{v}{(numeric) Uniform margin of copula}

\item{lambda}{(numeric) Dependence parameter of the Hüsler-Reiss copula}

\item{log}{(logical) If true returns log-Likelihood, default is False}
}
\value{
(numeric)
}
\description{
Fitting a bi-variate Hüsler-Reiss copula
}
\examples{
dat = rhr(1000, 1.2)
fit_hrc(dat[,1], dat[,2],1)
}
