% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{fit_hrc}
\alias{fit_hrc}
\title{Fitting a bi-variate H端sler-Reiss copula}
\usage{
fit_hrc(u, v, initial_est)
}
\arguments{
\item{u}{(numeric) Uniform margin of copula}

\item{v}{(numeric) Uniform margin of copula}

\item{lambda}{(numeric) Dependence parameter of the H端sler-Reiss copula}

\item{log}{(logical) If true returns log-Likelihood, default is False}
}
\value{
(list(estimate (numeric), ci (numeric))) List of 2 elements,
first element is the estimate of the H端sler-Reiss dependence parameter,
second element is the estimated 95% confidence interval from the Hessian.
}
\description{
Fitting a bi-variate H端sler-Reiss copula
}
\examples{
dat = rhr(1000, 1.2)
fit_hrc(dat[,1], dat[,2],1)
}
