% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_hr.R
\name{phr}
\alias{phr}
\title{Cumulative distribution function of the bi-variate Hüsler-Reiss copula}
\usage{
phr(u, v, theta)
}
\arguments{
\item{u}{(numeric) Uniform margin of copula}

\item{v}{(numeric) Uniform margin of copula}

\item{theta}{(numeric) dependence parameter of Hüsler-Reiss distribution}
}
\value{
(numeric)
}
\description{
Cumulative distribution function of the bi-variate Hüsler-Reiss copula
}
\examples{
dhr(u = c(0.1, 0.2, 0.4), v = c(0.6, 0.7, 0.2), lambda = 1)
}
