% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{qgev}
\alias{qgev}
\title{Quantile function of the generalised extreme value distribution (inverse cumulative distribution function)}
\usage{
qgev(x, loc, scale, shape)
}
\arguments{
\item{x}{(numeric) data}

\item{loc}{(numeric)}

\item{scale}{(numeric)}

\item{shape}{(numeric)}
}
\value{
(numeric)
}
\description{
Quantile function of the generalised extreme value distribution (inverse cumulative distribution function)
}
\examples{
qgev(x = 0.5, loc = 1, scale = 1, shape = 0)
}
